fix: use relative imports instead of path aliases
Path aliases depend on another step to resolve the path aliases in the compiled JS. The projects that use this library shouldn't be required to supply path aliases to make this library work. In tests we resolve paths with jest so it is OK to keep aliases in the testing tsconfig. See: https://github.com/microsoft/TypeScript/issues/26722